apps/om1-runtime/requirements.txt
# OM1 Dependencies
# Note: Install OM1 following official documentation at https://github.com/openmind/om1

# LLM APIs (Groq & OpenAI compatible)
openai==1.54.3

# HTTP Requests
requests==2.31.0
aiohttp==3.9.1

# Data Processing
numpy==1.26.2
pandas==2.1.4

# Sentiment Analysis
vaderSentiment==3.3.2

# Database
psycopg2-binary==2.9.9
redis==5.0.1

# Environment Variables
python-dotenv==1.0.0

# WebSocket
websockets==12.0

# JSON5 Configuration Parser
pyjson5==1.6.4

# Date/Time
python-dateutil==2.8.2

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
